requirements.txt
aniso8601==4.1.0
argon2-cffi==19.1.0
asn1crypto==0.24.0
astroid==1.6.5
autopep8==1.4.3
backports.functools-lru-cache==1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
idna==2.8
ipaddress==1.0.22
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
marshmallow==3.0.0rc3
mccabe==0.6.1
nexmo==2.3.0
peewee==3.8.2
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.7.1
pylint==1.9.4
pytz==2018.9
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
